Выполнение 23.1

+++ catalog/views.py
@@ -1,3 +1,5 @@
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponseForbidden
 from django.shortcuts import render,  get_object_or_404, redirect
 from catalog.models import Product
 from catalog.forms import ProductForm, VersionForm
@@ -82,16 +84,24 @@ def create_product(request):
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
-            form.save()
+            product = form.save(commit=False)
+
+            product.owner = request.user
+
+            product.save()
             return redirect('home')
     else:
         form = ProductForm()
 
     return render(request, 'catalog/create_product.html', {'form': form})
 
-
+@login_required
 def edit_product(request, pk):
     product = get_object_or_404(Product, pk=pk)
+
+
+    if request.user != product.owner:
+        return HttpResponseForbidden("У вас нет разрешения на редактирование этого продукта.")
 
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -103,9 +113,12 @@ def edit_product(request, pk):
 
     return render(request, 'catalog/edit_product.html', {'form': form, 'product': product})
 
-
+@login_required
 def delete_product(request, pk):
     product = get_object_or_404(Product, pk=pk)
+
+    if request.user != product.owner:
+        return HttpResponseForbidden("У вас нет разрешения на удаление этого продукта.")
 
     if request.method == 'POST':
         product.delete()
